Move all room functions out of pokertable
