+++ backend/src/game/PokerTable.ts
@@ -1,6 +1,5 @@
 // Internal
 import { Result, ResultError, ResultSuccess } from '@shared/Result';
-import { Rooms } from '../sockets/Rooms';
 import { Seat } from './Seat';
 import {
   PlayerAlreadySeatedError,
@@ -31,10 +30,6 @@ export class PokerTable {
   }
 
   public static createPokerTable(tableName: string, numberOfSeats: number): Result<PokerTable> {
-    const res = Rooms.createRoom(tableName);
-    if (res.error) {
-      return new ResultError(res.error);
-    }
     const newTable = new PokerTable(tableName, numberOfSeats);
     return new ResultSuccess(newTable);
   }
@@ -52,17 +47,6 @@ export class PokerTable {
         } else {
           seat.playerId = clientId;
           seat.isTaken = true;
-          const tableIsReady = this.checkTableReady();
-          if (tableIsReady) {
-            const event = 'start_game';
-            const payload = {
-              tableName: tableName,
-            };
-            const sendEvents = Rooms.sendEventToRoom('table_1', event, payload);
-            if (sendEvents.error) {
-              return Result.error(sendEvents.error);
-            }
-          }
           return Result.success();
         }
       }
@@ -96,7 +80,7 @@ export class PokerTable {
     return this.tableName;
   }
 
-  private checkTableReady(): boolean {
+  public checkTableReady(): boolean {
     for (const seat of this.seats) {
       if (seat.playerId == '') {
         return false;
